requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.8
asgiref==3.9.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.1.1